+++ CHANGES.rst
@@ -38,8 +38,10 @@ pending
 - |MAINTAIN| docker image based on python 3.10 <- 3.8
 - |CI|/CD: artifact uploading
 - |NEW| +1 named color
+- |REFACTOR| `TemplateEngine`
+- |TESTS| `template`
 
-.. <@pending:814f67e>
+.. <@pending:bff2eaa>
 .. ^ blank line before should be kept